Split out getJetpackUpgradeUrlIfPremiumTheme
